More kind checking, issue 944
