+++ core/src/main/scala/org/bykn/bosatsu/Package.scala
@@ -1,7 +1,7 @@
 package org.bykn.bosatsu
 
 import cats.{Functor, Parallel}
-import cats.data.{Chain, Ior, ValidatedNel, Validated, NonEmptyList, Writer, NonEmptyMap}
+import cats.data.{Ior, ValidatedNel, Validated, NonEmptyList}
 import cats.implicits._
 import cats.parse.{Parser0 => P0, Parser => P}
 import org.typelevel.paiges.{Doc, Document}
@@ -11,7 +11,6 @@ import rankn._
 import Parser.{spaces, Combinators}
 
 import FixType.Fix
-import LocationMap.Colorize
 /**
  * Represents a package over its life-cycle: from parsed to resolved to inferred
  */
@@ -349,390 +348,4 @@ object Package {
         val lm = LocationMap(Predef.predefString)
         sys.error(s"couldn't parse predef: ${lm.showContext(idx, 2, LocationMap.Colorize.None)} with errs: ${err}")
     }
-}
-
-
-sealed abstract class PackageError {
-  def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize): String
-
-  protected def getMapSrc(sourceMap: Map[PackageName, (LocationMap, String)], pack: PackageName): (LocationMap, String) =
-    sourceMap.get(pack) match {
-      case None => (LocationMap(""), "<unknown source>")
-      case Some(found) => found
-    }
-}
-
-object PackageError {
-  def showTypes(pack: PackageName, tpes: List[Type]): Map[Type, Doc] = {
-    // TODO: we should use the imports in each package to talk about
-    // types in ways that are local to that package
-    require(pack ne null)
-    tpes
-      .iterator
-      .map { t =>
-        (t, Type.fullyResolvedDocument.document(t))
-      }
-      .toMap
-  }
-
-  def nearest[A](ident: Identifier, existing: Map[Identifier, A], count: Int): List[(Identifier, A)] =
-    existing
-      .iterator
-      .map { case (i, a) =>
-        val d = EditDistance.string(ident.asString, i.asString)
-        (i, d, a)
-      }
-      .filter(_._2 < ident.asString.length) // don't show things that require total edits
-      .toList
-      .sortBy { case (_, d, _) => d }
-      .distinct
-      .take(count)
-      .map { case (i, _, a) => (i, a) }
-
-  case class UnknownExport[A](ex: ExportedName[A],
-    in: PackageName,
-    lets: List[(Identifier.Bindable, RecursionKind, TypedExpr[Declaration])]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, in)
-      val header =
-        s"in $sourceName unknown export ${ex.name.sourceCodeRepr}"
-      val candidateMap: Map[Identifier, Region] =
-        lets.map { case (n, _, expr) => (n, HasRegion.region(expr)) }.toMap
-      val candidates =
-        nearest(ex.name, candidateMap, 3)
-          .map { case (n, r) =>
-            val pos = lm.toLineCol(r.start).map { case (l, c) => s" at line: ${l + 1}, column: ${c + 1}" }.getOrElse("")
-            s"${n.asString}$pos"
-          }
-      val candstr = candidates.mkString("\n\t", "\n\t", "\n")
-      val suggestion =
-        if (candidates.nonEmpty) "\n" + s"perhaps you meant:$candstr"
-        else ""
-      header + suggestion
-    }
-  }
-
-  case class PrivateTypeEscape[A](ex: ExportedName[A],
-    exType: Type,
-    in: PackageName,
-    privateType: Type.Const) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (_, sourceName) = getMapSrc(sourceMap, in)
-      val pt = Type.TyConst(privateType)
-      val tpeMap = showTypes(in, exType :: pt :: Nil)
-      val first = s"in $sourceName export ${ex.name.sourceCodeRepr} of type ${tpeMap(exType).render(80)}"
-      if (exType == pt) {
-        s"$first has an unexported (private) type."
-      }
-      else {
-        s"$first references an unexported (private) type ${tpeMap(pt).render(80)}."
-      }
-    }
-  }
-
-  case class UnknownImportPackage[A, B, C](pack: PackageName, from: Package[PackageName, A, B, C]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (_, sourceName) = getMapSrc(sourceMap, from.name)
-      s"in $sourceName package ${from.name.asString} imports unknown package ${pack.asString}"
-    }
-  }
-
-  case class DuplicatedImport(duplicates: NonEmptyList[(PackageName, ImportedName[Unit])]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) =
-      duplicates.sortBy(_._2.localName)
-        .toList
-        .iterator
-        .map { case (pack, imp) =>
-          val (_, sourceName) = getMapSrc(sourceMap, pack)
-          s"duplicate import in $sourceName package ${pack.asString} imports ${imp.originalName.sourceCodeRepr} as ${imp.localName.sourceCodeRepr}"
-        }
-        .mkString("\n")
-  }
-
-  // We could check if we forgot to export the name in the package and give that error
-  case class UnknownImportName[A, B](
-    in: PackageName,
-    importing: Package.Inferred,
-    iname: ImportedName[A],
-    exports: List[ExportedName[B]]) extends PackageError {
-      def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-        val ls = importing
-          .program
-          .lets
-
-        val (_, sourceName) = getMapSrc(sourceMap, in)
-        val letMap = ls.iterator.map { case (n, _, _) => (n: Identifier, ()) }.toMap
-        letMap
-          .get(iname.originalName) match {
-            case Some(_) =>
-              s"in $sourceName package: ${importing.name.asString} has ${iname.originalName.sourceCodeRepr} but it is not exported. Add to exports"
-            case None =>
-              val near = nearest(iname.originalName, letMap, 3)
-                .map { case (n, _) => n.sourceCodeRepr }
-                .mkString(" Nearest: ", ", ", "")
-              s"in $sourceName package: ${importing.name.asString} does not have name ${iname.originalName.sourceCodeRepr}.$near"
-          }
-      }
-    }
-
-  case class UnknownImportFromInterface[A, B](
-    in: PackageName,
-    importing: Package.Interface,
-    iname: ImportedName[A],
-    exports: List[ExportedName[B]]) extends PackageError {
-      def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-
-        val (_, sourceName) = getMapSrc(sourceMap, in)
-        val exportMap = importing.exports.map { e => (e.name, ()) }.toMap
-        val near = nearest(iname.originalName, exportMap, 3)
-          .map { case (n, _) => n.asString }
-          .mkString(" Nearest: ", ", ", "")
-        s"in $sourceName package: ${importing.name} does not have name ${iname.originalName}.$near"
-      }
-    }
-
-  case class CircularDependency[A, B, C](from: Package[PackageName, A, B, C], path: NonEmptyList[PackageName]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val packs = from.name :: (path.toList)
-      val msg = packs.map { p =>
-        val (_, src) = getMapSrc(sourceMap, p)
-        s"${p.asString} in $src"
-      }
-      val tab = "\n\t"
-      s"circular package dependency:\n${msg.mkString(tab)}"
-    }
-  }
-
-  case class CircularType[A](from: PackageName, path: NonEmptyList[rankn.DefinedType[A]]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      s"circular types in ${from.asString} " + path.toList.reverse.map(_.name.ident.asString).mkString(" -> ")
-    }
-  }
-
-  case class VarianceInferenceFailure(from: PackageName, failed: NonEmptyList[rankn.DefinedType[Unit]]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      s"failed to infer variance in ${from.asString} of " + failed.toList.map(_.name.ident.asString).sorted.mkString(", ")
-    }
-  }
-
-  case class TypeErrorIn(tpeErr: Infer.Error, pack: PackageName) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, pack)
-      val teMessage = tpeErr match {
-        case Infer.Error.NotUnifiable(t0, t1, r0, r1) =>
-          val context0 =
-            if (r0 == r1) Doc.space // sometimes the region of the error is the same on right and left
-            else {
-              val m = lm.showRegion(r0, 2, errColor).getOrElse(Doc.str(r0)) // we should highlight the whole region
-              Doc.hardLine + m + Doc.hardLine
-            }
-          val context1 =
-            lm.showRegion(r1, 2, errColor).getOrElse(Doc.str(r1)) // we should highlight the whole region
-
-          val fnHint =
-            (t0, t1) match {
-              case (Type.Fun(_, _), Type.Fun(_, _)) =>
-                // both are functions
-                Doc.empty
-              case (Type.Fun(_, _), _) | (_, Type.Fun(_, _)) =>
-                Doc.text("hint: this often happens when you apply the wrong number of arguments to a function.") + Doc.hardLine
-              case _ =>
-                Doc.empty
-            }
-
-          val tmap = showTypes(pack, List(t0, t1))
-          val doc = Doc.text("type error: expected type ") + tmap(t0) +
-            context0 + Doc.text("to be the same as type ") + tmap(t1) +
-            Doc.hardLine + fnHint + context1
-
-          doc.render(80)
-        case Infer.Error.VarNotInScope((_, name), scope, region) =>
-          val ctx = lm.showRegion(region, 2, errColor).getOrElse(Doc.str(region))
-          val candidates: List[String] =
-            nearest(name, scope.map { case ((_, n), _) => (n, ()) }, 3)
-              .map { case (n, _) => n.asString }
-
-          val cmessage =
-            if (candidates.nonEmpty) candidates.mkString("\nClosest: ", ", ", ".\n")
-            else ""
-          val qname = "\"" + name.sourceCodeRepr + "\""
-          (Doc.text("name ") + Doc.text(qname) + Doc.text(" unknown.") + Doc.text(cmessage) + Doc.hardLine +
-            ctx).render(80)
-        case Infer.Error.SubsumptionCheckFailure(t0, t1, r0, r1, _) =>
-          val context0 =
-            if (r0 == r1) Doc.space // sometimes the region of the error is the same on right and left
-            else {
-              val m = lm.showRegion(r0, 2, errColor).getOrElse(Doc.str(r0)) // we should highlight the whole region
-              Doc.hardLine + m + Doc.hardLine
-            }
-          val context1 =
-            lm.showRegion(r1, 2, errColor).getOrElse(Doc.str(r1)) // we should highlight the whole region
-
-          val tmap = showTypes(pack, List(t0, t1))
-          val doc = Doc.text("type ") + tmap(t0) + context0 +
-            Doc.text("does not subsume type ") + tmap(t1) + Doc.hardLine +
-            context1
-
-          doc.render(80)
-        case uc@Infer.Error.UnknownConstructor((_, n), region, _) =>
-          val near = nearest(n, uc.knownConstructors.map { case (_, n) => (n, ()) }.toMap, 3)
-            .map { case (n, _) => n.asString }
-
-          val nearStr =
-            if (near.isEmpty) ""
-            else near.mkString(", nearest: ", ", ", "")
-
-          val context =
-            lm.showRegion(region, 2, errColor).getOrElse(Doc.str(region)) // we should highlight the whole region
-
-          val doc = Doc.text("unknown constructor ") + Doc.text(n.asString) +
-            Doc.text(nearStr) + Doc.hardLine + context
-          doc.render(80)
-        case Infer.Error.KindInvariantViolation(tpe, right, region, mess) =>
-          val applied = Type.TyApply(tpe, right)
-          val tmap = showTypes(pack, applied :: Nil)
-          val context =
-            lm.showRegion(region, 2, errColor).getOrElse(Doc.str(region)) // we should highlight the whole region
-          val doc = Doc.text("kind error: ") + Doc.text(mess) + Doc.text(" for type the left of") + tmap(applied) + Doc.hardLine +
-            context
-
-          doc.render(80)
-        case err => err.message
-      }
-      // TODO use the sourceMap/regiouns in Infer.Error
-      s"in file: $sourceName, package ${pack.asString}, $teMessage"
-    }
-  }
-
-  case class SourceConverterErrorIn(err: SourceConverter.Error, pack: PackageName) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, pack)
-      val msg = {
-        val context =
-          lm.showRegion(err.region, 2, errColor)
-            .getOrElse(Doc.str(err.region)) // we should highlight the whole region
-
-        Doc.text(err.message) + Doc.hardLine + context
-      }
-      val doc = Doc.text("in file: ") + Doc.text(sourceName) + Doc.text(", package ") + Doc.text(pack.asString) +
-        Doc.text(", ") + msg
-
-      doc.render(80)
-    }
-  }
-
-  case class TotalityCheckError(pack: PackageName, err: TotalityCheck.ExprError[Declaration]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, pack)
-      val region = err.matchExpr.tag.region
-      val context1 =
-        lm.showRegion(region, 2, errColor).getOrElse(Doc.str(region)) // we should highlight the whole region
-      val teMessage = err match {
-        case TotalityCheck.NonTotalMatch(_, missing) =>
-          val allTypes = missing.traverse(_.traverseType { t => Writer(Chain.one(t), ()) })
-            .run._1.toList.distinct
-          val showT = showTypes(pack, allTypes)
-
-          val doc = Pattern.compiledDocument(Document.instance[Type] { t =>
-            showT(t)
-          })
-
-          Doc.text("non-total match, missing: ") +
-            (Doc.intercalate(Doc.char(',') + Doc.lineOrSpace,
-              missing.toList.map(doc.document(_))))
-        case TotalityCheck.UnreachableBranches(_, unreachableBranches) =>
-          val allTypes = unreachableBranches.traverse(_.traverseType { t => Writer(Chain.one(t), ()) })
-            .run._1.toList.distinct
-          val showT = showTypes(pack, allTypes)
-
-          val doc = Pattern.compiledDocument(Document.instance[Type] { t =>
-            showT(t)
-          })
-
-          Doc.text("unreachable branches: ") +
-            (Doc.intercalate(Doc.char(',') + Doc.lineOrSpace,
-              unreachableBranches.toList.map(doc.document(_))))
-        case TotalityCheck.InvalidPattern(_, err) =>
-          import TotalityCheck._
-          err match {
-            case ArityMismatch((_, n), _, _, exp, found) =>
-              Doc.text(s"arity mismatch: ${n.asString} expected $exp parameters, found $found")
-            case UnknownConstructor((_, n), _, _) =>
-              Doc.text(s"unknown constructor: ${n.asString}")
-            case InvalidStrPat(pat, _) =>
-              Doc.text(s"invalid string pattern: ") +
-                Document[Pattern.Parsed].document(pat) +
-                Doc.text(" (adjacent bindings aren't allowed)")
-            case MultipleSplicesInPattern(_, _) =>
-              // TODO: get printing of compiled patterns working well
-              //val docp = Document[Pattern.Parsed].document(Pattern.ListPat(pat)) +
-              Doc.text("multiple splices in pattern, only one per match allowed")
-          }
-      }
-      // TODO use the sourceMap/regions in Infer.Error
-      val doc = Doc.text(s"in file: $sourceName, package ${pack.asString}") + Doc.hardLine +
-        context1 + Doc.hardLine + teMessage
-
-      doc.render(80)
-    }
-  }
-
-  case class UnusedLetError(pack: PackageName, errs: NonEmptyList[(Identifier.Bindable, Region)]) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, pack)
-      val docs = errs
-        .sortBy(_._2)
-        .map { case (bn, region) =>
-          val rdoc = lm.showRegion(region, 2, errColor).getOrElse(Doc.str(region)) // we should highlight the whole region
-          val message = Doc.text("unused let binding: " + bn.sourceCodeRepr)
-          message + Doc.hardLine + rdoc
-        }
-
-      val packDoc = Doc.text(s"in file: $sourceName, package ${pack.asString}:")
-      val line2 = Doc.hardLine + Doc.hardLine
-      (packDoc + (line2 + Doc.intercalate(line2, docs.toList)).nested(2)).render(80)
-    }
-  }
-
-  case class RecursionError(pack: PackageName, err: DefRecursionCheck.RecursionError) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val (lm, sourceName) = getMapSrc(sourceMap, pack)
-      val ctx = lm.showRegion(err.region, 2, errColor)
-        .getOrElse(Doc.str(err.region)) // we should highlight the whole region
-      val errMessage = err.message
-      // TODO use the sourceMap/regions in RecursionError
-      val doc = Doc.text(s"in file: $sourceName, package ${pack.asString}, $errMessage") +
-        Doc.hardLine + ctx + Doc.hardLine
-
-      doc.render(80)
-    }
-  }
-
-  case class DuplicatedPackageError[A](dups: NonEmptyMap[PackageName, ((A, Package.Parsed), NonEmptyList[(A, Package.Parsed)])], show: A => String) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      val packDoc = Doc.text("package ")
-      val dupInDoc = Doc.text(" duplicated in ")
-      val dupMessages = dups
-        .toSortedMap
-        .map { case (pname, (one, nelist)) =>
-          val dupsrcs = Doc.intercalate(Doc.comma + Doc.lineOrSpace,
-            (one :: nelist.toList)
-              .map { case (s, _) => show(s) }
-              .sorted
-              .map(Doc.text(_))
-            )
-            .nested(4)
-          packDoc + Doc.text(pname.asString) + dupInDoc + dupsrcs
-        }
-
-      Doc.intercalate(Doc.line, dupMessages).render(80)
-    }
-  }
-
-  case class KindInferenceError(pack: PackageName, kindError: KindFormula.Error) extends PackageError {
-    def message(sourceMap: Map[PackageName, (LocationMap, String)], errColor: Colorize) = {
-      // TODO actually make this look sane
-      s"in package ${pack.asString}: KindError: $kindError"
-    }
-  }
 }
